Adding index html with facebook SDK
I have the html file with the facebook login with JavaScript SDK and the button plugin itself. Instructions stated that the code is to be kept immediately after the opening body tag, but we can include the code for the facebook plugin itself wherever we please. We also have the option to include the SDK in another page, but this is what I have so far.